added invalidpage component, updated about page
